+++ tests/test_parser.py
@@ -16,82 +16,23 @@
 """ unit tests for xmldiff.
 """
 
-import unittest
-from xmldiff.objects import *
+import os
+import mock
 import six
-import sys
+from xmldiff.parser import SaxHandler
+from xmldiff.input import tree_from_stream
+from xml.sax import make_parser
+
+from xmldiff.objects import NT_ROOT, NT_NODE, NT_ATTN, NT_ATTV, \
+    NT_TEXT, NT_COMM, N_TYPE, N_NAME, N_VALUE, N_CHILDS, N_PARENT, N_ISSUE, \
+    N_XNUM, NSIZE, A_DESC, A_N1, A_N2, FALSE, TRUE
 
 
-class Tree2TreeKnownValues(unittest.TestCase):
-    """
-    This class check that source_vector functions give known result
-    with known input
-    """
+HERE = os.path.dirname(__file__)
 
-    DATA = six.StringIO("""
-  <memory>
-  <mailbox path="/var/spool/mail/syt"/>
-  <email_addr mine="yes">syt@logilab.org</email_addr>
-  <server-socket1 port="7776" recipe="pia.PDA"/>
-   <bookmark-file type="netscape">/home/syt/.netscape/bookmarks.html</bookmark-file>
-   <!-- proxy -->
-  <server-socket2 port="7777" recipe="proxy.Web proxy"/>
-  <junkbuster-method value="18" />
-  <spoken-languages>
-   <language1 name="english" code="en" />
-   <language2 name="francais" code="fr" />
-  </spoken-languages>
-</memory>
-""")
-    from xmldiff.parser import SaxHandler
-    from xml.sax import make_parser
-    dh = SaxHandler(1, 1)
-    parser = make_parser()
-    parser.setContentHandler(dh)
-    parser.parse(DATA)
-    xmltree1 = dh.get_tree()
 
-    DATA = six.StringIO("""
-   <memory>
-  <mailbox path="/var/spooool/mail/syt"/>
-  <email_addr mine="yes">syt@logab.org</email_addr>
-  <server-socket1 port="7776" recipe="pia.PDA"/>
-  <bookmark-file>/home/syt/.netscape/bookmarks.html</bookmark-file>
-  <!-- proxy -->
-  <junkbuster-method value="18" />
-  <server-socket2 port="7777" recipe="proxy.Web proxy"/>
-  <spoken-languages>
-   <language1 code="en" name="english"/>
-   <language2 name="francais" code="fr" />
-   <language3 name="martien" code="ma"/>
-  </spoken-languages>
-</memory>
-""")
-    dh.__init__(1, 1)
-    parser.parse(DATA)
-    xmltree2 = dh.get_tree()
-
-    # those variables may change if the "DATA" string change
-    # this is the post order numbers
-    HKNOWN_VALUES = {
-        'N/a[0]': 14,
-        'a/a[0]': 1,
-        'N/a[0]/b[0]': 3,
-        'b/a[0]/b[0]': 2,
-        'N/a[0]/c[0]': 5,
-        'c/a[0]/c[0]': 4,
-        'N/a[0]/d[0]': 13,
-        'd/a[0]/d[0]': 6,
-        'N/a[0]/d[0]/e[0]': 10,
-        'e/a[0]/d[0]/e[0]': 7,
-        'N/a[0]/d[0]/e[0]/h[0]': 9,
-        'h/a[0]/d[0]/e[0]/h[0]': 8,
-        'N/a[0]/d[0]/f[0]': 12,
-        'f/a[0]/d[0]/f[0]': 11
-    }
-
-    dh.__init__(1, 1)
-    parser.parse(six.StringIO("""
+def test_tree_from_stream_simple():
+    stream = six.StringIO("""
     <a>
       <b/>
       <c>
@@ -104,49 +45,85 @@ class Tree2TreeKnownValues(unittest.TestCase):
         </f>
       </d>
     </a>
-    """))
-    tree1 = dh.get_tree()
+    """)
+    tree = tree_from_stream(stream)
+    expected = [
+        6,
+        '/',
+        '',
+        [[1,
+          u'a',
+          u'a',
+          [[1, u'b', u'b', [], mock.ANY, 0, 1],
+           [1, u'c', u'c', [], mock.ANY, 0, 1],
+           [1,
+            u'd',
+            u'd',
+            [[1,
+              u'e',
+              u'e',
+              [[1,
+                u'h',
+                u'h',
+                [],
+                mock.ANY,
+                0,
+                1]],
+              mock.ANY,
+              1,
+              1],
+             [1,
+              u'f',
+              u'f',
+              [],
+              mock.ANY,
+              0,
+              1]],
+            mock.ANY,
+            3,
+            1]],
+          mock.ANY,
+          6,
+          1]],
+        None,
+        7,
+        0]
+    assert tree == expected
 
-    dh.__init__(1, 1)
-    parser.parse(six.StringIO("""
-    <a>
-      <c/>
-      <b>
-      </b>
-      <d>
-        <e>
-          <h/>
-        </e>
-        <f><j/>
-        </f>
-      </d>
-    </a>
-    """))
-    tree2 = dh.get_tree()
 
-    def setUp(self):
-        """ called before each test from this class """
-        self.nl1, self.nl2 = [], []
+def test_tree_from_stream():
+    fname = os.path.join(HERE, 'data', 'parse', '1.xml')
+    with open(fname, 'r') as fhandle:
+        tree = tree_from_stream(fhandle)
+        # lets not dump the whole tree
+        assert len(tree[N_CHILDS]) == 1
 
 
-def suite():
-    """return the unitest suite"""
-    loader = unittest.TestLoader()
-    module = sys.modules[__name__]
-    if __name__ == '__main__' and len(sys.argv) > 1:
-        return loader.loadTestsFromNames(sys.argv[1:], module)
-    return loader.loadTestsFromModule(module)
+def test_tree_from_stream_utf8():
+    fname = os.path.join(HERE, 'data', 'parse', 'utf8.xml')
+    with open(fname, 'r') as fhandle:
+        tree = tree_from_stream(fhandle)
+        type_node = tree[N_CHILDS][0][N_CHILDS][0][N_CHILDS][0][N_CHILDS][0]
+        text_node = tree[N_CHILDS][0][N_CHILDS][0][N_CHILDS][1][N_CHILDS][0]
+        assert type_node[N_VALUE] == u'\xf6\xfc'
+        assert text_node[N_VALUE] == u'\xe9\xe1\u03a9'
 
 
-def Run(runner=None):
-    """run tests"""
-    testsuite = suite()
-    if runner is None:
-        runner = unittest.TextTestRunner()
-        # uncomment next line to write tests results in a file
-        # runner.__init__(open('tests.log','w+'))
-    return runner.run(testsuite)
+def test_tree_from_stream_utf16():
+    fname = os.path.join(HERE, 'data', 'parse', 'utf16.xml')
+    with open(fname, 'r') as fhandle:
+        tree = tree_from_stream(fhandle)
+        type_node = tree[N_CHILDS][0][N_CHILDS][0][N_CHILDS][0][N_CHILDS][0]
+        text_node = tree[N_CHILDS][0][N_CHILDS][0][N_CHILDS][1][N_CHILDS][0]
+        assert type_node[N_VALUE] == u'\xf6\xfc'
+        assert text_node[N_VALUE] == u'\xe9\xe1\u03a9'
 
 
-if __name__ == '__main__':
-    Run()
+def test_tree_from_stream_iso():
+    fname = os.path.join(HERE, 'data', 'parse', 'iso.xml')
+    with open(fname, 'r') as fhandle:
+        tree = tree_from_stream(fhandle)
+        type_node = tree[N_CHILDS][0][N_CHILDS][0][N_CHILDS][0][N_CHILDS][0]
+        text_node = tree[N_CHILDS][0][N_CHILDS][0][N_CHILDS][1][N_CHILDS][0]
+        assert type_node[N_VALUE] == u'\xf6\xfc'
+        assert text_node[N_VALUE] == u'\xe9\xe1'